# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: Deprecated public venue directory
      deprecated: true
      responses:
        "404":
          description: Directory disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a new venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueBody"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /venues/{slug}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get a venue by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Venue found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenuePublicResponse"
        "404":
          description: Venue not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update venue profile (owner session)
      description: |
        Owner-only. Updates the public-facing venue profile fields. The slug
        cannot be changed (it is the public URL handle and would break QR codes
        and shared links). Owner email updates require an active owner session.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueBody"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Venue not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /venues/{slug}/dashboard:
    get:
      operationId: getVenueDashboard
      tags: [venues]
      summary: Get venue dashboard (owner session)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDashboardResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Venue not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /venues/{slug}/stats:
    get:
      operationId: getVenueStats
      tags: [venues]
      summary: Get venue statistics
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Venue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueStatsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /venues/{slug}/media:
    get:
      operationId: listVenueMedia
      tags: [venues]
      summary: List venue media (owner session)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of venue media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListVenueMediaResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Venue not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: addVenueMedia
      tags: [venues]
      summary: Add a media item to the venue (owner session)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVenueMediaBody"
      responses:
        "201":
          description: Media added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueMediaItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /venues/{slug}/media/{mediaId}:
    delete:
      operationId: deleteVenueMedia
      tags: [venues]
      summary: Delete a venue media item (owner session)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Media not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CreateVenueBody:
      type: object
      required: [name, slug, ownerEmail]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
          pattern: "^[a-z0-9-]+$"
        tagline:
          type: string
        description:
          type: string
        ownerEmail:
          type: string
          format: email
          description: Email of the venue owner for secure login links and notifications.
        contactEmail:
          type: string
          format: email
          description: Optional public inquiry email shown to couples.
        contactPhone:
          type: string
          description: Optional public phone number shown to couples.
        websiteUrl:
          type: string
          format: uri
          description: Optional public venue website URL.
        bookingUrl:
          type: string
          format: uri
          description: Optional public tour booking URL.
    VenueResponse:
      type: object
      required: [id, name, slug, plan, creditsBalance, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        ownerEmail:
          type: string
          format: email
          nullable: true
          description: Owner contact email (only returned on owner-session protected responses).
        contactEmail:
          type: string
          nullable: true
          description: Public inquiry email shown to couples.
        contactPhone:
          type: string
          nullable: true
          description: Public phone number shown to couples.
        websiteUrl:
          type: string
          nullable: true
          description: Public venue website URL.
        bookingUrl:
          type: string
          nullable: true
          description: Public tour booking URL.
        organizationId:
          type: integer
          nullable: true
          description: Billing organization that owns this venue.
        plan:
          type: string
          enum: [trial, starter, growth, none]
          description: Organization plan (billing lives on the organization).
        creditsBalance:
          type: integer
          description: Organization credit balance (shared across the org's venues).
        billingPeriodEnd:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    UpdateVenueBody:
      type: object
      description: |
        All fields optional. Only the provided fields are updated.
        Slug is intentionally not editable to keep public URLs stable.
      properties:
        name:
          type: string
          minLength: 1
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        ownerEmail:
          type: string
          format: email
          description: Owner email for secure login links and notifications. Cannot be cleared.
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        bookingUrl:
          type: string
          nullable: true
...
    VenuePublicResponse:
      type: object
      required: [id, name, slug, media, isReady]
      description: Couple-facing venue profile. Never exposes billing or credit balances.
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
          description: Public inquiry email shown to couples.
        contactPhone:
          type: string
          nullable: true
          description: Public phone number shown to couples.
        websiteUrl:
          type: string
          nullable: true
          description: Public venue website URL.
        bookingUrl:
          type: string
          nullable: true
          description: Public tour booking URL.
        createdAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/VenueMediaItem"
        isReady:
          type: boolean
          description: True when the venue has at least five uploaded photos and complete required coverage.
        uploadToken:
          type: string
          description: Short-lived venue-scoped token required for couple photo upload URL requests.
...
    VenueMediaItem:
      type: object
      required: [id, venueId, objectKey, coverage, displayOrder, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        objectKey:
          type: string
        coverage:
          $ref: "#/components/schemas/VenueMediaCoverage"
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    VenueMediaCoverage:
      type: string
      enum: [exterior, ceremony, reception, detail, natural_light]
      description: Coverage role of the venue reference photo for scene-specific generation.
    ListVenueMediaResponse:
      type: object
      required: [media]
      properties:
        media:
          type: array
          items:
            $ref: "#/components/schemas/VenueMediaItem"
    AddVenueMediaBody:
      type: object
      required: [objectKey, coverage]
      properties:
        objectKey:
          type: string
        coverage:
          $ref: "#/components/schemas/VenueMediaCoverage"
        displayOrder:
          type: integer
    VenueDashboardResponse:
      type: object
      required: [venue, sessions]
      properties:
        venue:
          $ref: "#/components/schemas/VenueResponse"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
    VenueStatsResponse:
      type: object
      required: [totalSessions, readySessions, processingSessions, failedSessions, avgCompletionSeconds, totalEstimatedCostUsd]
      properties:
        totalSessions:
          type: integer
        readySessions:
          type: integer
        processingSessions:
          type: integer
        failedSessions:
          type: integer
        avgCompletionSeconds:
          type: number
          nullable: true
        totalEstimatedCostUsd:
          type: number
          description: Sum of estimated USD costs across all sessions for this venue.
...